--- .editorconfig
@@ -1,15 +1,0 @@
-root = true
-
-[*]
-indent_style = tab
-indent_size = 4
-end_of_line = lf
-insert_final_newline = true
-
-[{*.md}]
-indent_style = space
-indent_size = 4
-
-[{*.yaml,*.yml,package.json}]
-indent_style = space
-indent_size = 2
