--- .eslintrc.js
@@ -1,9 +1,0 @@
-module.exports = {
-	root: true,
-	extends: [
-		'bluedrop',
-		'bluedrop/config/ecmascript-9',
-		'bluedrop/config/console',
-		'bluedrop/config/node',
-	],
-};
