+++ CopyPastaBot.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27428.2015
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9092AA53-FB77-4645-B42D-1CCCA6BD08BD}") = "CopyPastaBot", "CopyPastaBot\CopyPastaBot.njsproj", "{B124CE67-5361-4231-9D8C-504DF2D63312}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B124CE67-5361-4231-9D8C-504DF2D63312}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B124CE67-5361-4231-9D8C-504DF2D63312}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B124CE67-5361-4231-9D8C-504DF2D63312}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B124CE67-5361-4231-9D8C-504DF2D63312}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A115669E-C5FA-446D-B91E-E14D4DF07E3F}
+	EndGlobalSection
+EndGlobal
