+++ CopyPastaBot/Dockerfile
