+++ CopyPastaBot/commands.json
@@ -1,0 +1,44 @@
+[
+      {
+        "Command": "help",
+        "HandlerFile": "helpCommand.js",
+        "HandlerIndex": 0,
+        "Description": "This gives the help command"
+      },
+      {
+        "Command": "list",
+        "HandlerFile": "listCommand.js",
+        "HandlerIndex": 1,
+        "Description": "returns the list of indexed copypasta's."
+      },
+      {
+        "Command": "copypasta",
+        "HandlerFile": "copypastaCommand.js",
+        "HandlerIndex": 3,
+        "Description": "providing a valid copypasta id it replies with that copypasta."
+      },
+      {
+        "Command": "wisdom",
+        "HandlerFile": "randomCommand.js",
+        "HandlerIndex": 2,
+        "Description": "gives a random copypasta."
+      },
+      {
+        "Command": "voice",
+        "HandlerFile": "voiceCommand.js",
+        "HandlerIndex": 4,
+        "Description": "broadcasts a copypasta."
+      },
+      {
+        "Command": "skip",
+        "HandlerFile": "voiceCommand.js",
+        "HandlerIndex": 4,
+        "Description": "skips the current copypasta for the next one."
+      },
+      {
+        "Command": "stop",
+        "HandlerFile": "voiceCommand.js",
+        "HandlerIndex": 4,
+        "Description": "stops the broadcasts of a copypasta."
+      }
+    ]
