+++ CopyPastaBot/docker-entry.sh
