+++ CopyPastaBot/package.json
@@ -7,25 +7,25 @@
     "name": "Stephen"
   },
   "dependencies": {
-    "@google-cloud/storage": "^3.0.2",
-    "@google-cloud/text-to-speech": "^1.1.2",
-    "discord.js": "^11.5.1",
+    "@google-cloud/storage": "^2.5.0",
+    "@google-cloud/text-to-speech": "^0.5.1",
+    "discord.js": "^11.4.2",
     "ffmpeg": "0.0.4",
     "ffmpeg-static": "^2.4.0",
     "forever": "^1.0.0",
-    "fs-extra": "^8.1.0",
+    "fs": "0.0.1-security",
+    "http": "0.0.0",
     "opusscript": "0.0.6",
     "prism-media": "^1.0.2",
-    "snoowrap": "^1.17.0",
-    "sqlite3": "^4.0.9",
+    "snoowrap": "^1.15.2",
+    "sqlite3": "^4.0.6",
     "streamifier": "^0.1.1"
   },
   "devDependencies": {
     "eslint": "^6.0.1",
     "eslint-config-bluedrop": "^6.0.0",
     "eslint-plugin-filenames": "^1.3.2",
-    "eslint-plugin-jsdoc": "^10.0.3",
-    "eslint-plugin-node": "^9.1.0",
+    "eslint-plugin-jsdoc": "^8.7.0",
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-security": "^1.4.0"
   }
