+++ CopyPastaBot/voiceCommand.js
@@ -1,0 +1,146 @@
+﻿const fs = require('fs');
+const util = require('util');
+// Imports the Google Cloud client library
+const textToSpeech = require('@google-cloud/text-to-speech');
+const path = require('path');
+const http = require('http');
+const streamifier = require('streamifier');
+const prism = require('prism-media');
+
+let database;
+let r;
+let client;
+let queued = [];
+
+module.exports = {
+
+    init(data) {
+        database = data.database;
+        r = data.r;
+        client = data.client;
+    },
+
+    async CommandHandler(message, cmd, args) {
+
+        let vc = message.author.lastMessage.member.voiceChannelID;
+
+        if (vc === null || vc === undefined) {
+            message.reply("You have to be in a voice channel to suffer my pain!!", { tts: true });
+            return;
+        }
+
+        //stop and skip commands
+        if (cmd === 'stop') {
+            queued = [];
+            if (client.voiceConnections.get(message.guild.id) !== undefined) {
+                client.voiceConnections.get(message.guild.id).disconnect();
+            }
+            return;
+        } else if (cmd === 'skip') {
+            if (client.voiceConnections.get(message.guild.id) !== undefined) {
+                client.voiceConnections.get(message.guild.id).disconnect();
+                if (queued.length !== 0) {
+                    let next = queued.pop();
+                    this.playText(next.text, next.vc);
+                }
+            }
+            return;
+        }
+
+
+        //doing database checks
+
+        let exit = false;
+        let sub;
+        let text;
+        switch (args[0]) {
+            case 'post':
+                sub = await r.getSubmission(args[1]);
+                text = await sub.selftext;
+                //some edge case filtering
+                if (text.length === 0) {
+                    text = await sub.title;
+                }
+                let in_db = await database.checkPost(args[1]);
+
+
+                if (in_db === undefined) {
+                    database.addPost(args[1], await sub.title);
+                }
+                break;
+            case 'comment':
+                text = await r.getComment(args[1]).body;
+                break;
+            default:
+                message.reply("Your command has an issue, I can't suffer now", { tts: true });
+                exit = true;
+                break;
+        }
+        if (exit) {
+            return;
+        }
+        if (client.voiceConnections.get(message.guild.id) !== undefined) {
+            queued.push({ text: text, vc: vc });
+        } else {
+            this.playText(text, vc);
+        }
+
+    },
+
+    playTextTest(text, vc) {
+        http.get(`http://127.0.0.1:8080/speech?text="hello%20World"&encoding=opus`,async response => {
+            if (!err) {
+        // Performs the Text-to-Speech request
+        const [response] = await c.synthesizeSpeech(request);
+        const strm = streamifier.createReadStream(response.audioContent);
+        const audio = strm.pipe(new prism.opus.OggDemuxer());
+
+        let channel = client.channels.find(c => c.id === vc);
+        await channel.join().then(async (connection) => {
+            connection.playOpusStream(audio).on('end', () => {
+                if (queued.length !== 0) {
+                    let next = queued.pop();
+                    module.exports.playText(next.text, next.vc);
+                } else {
+                    channel.leave();
+                }
+            });
+        }).catch(err => console.log(err));
+            } else {
+                console.log(err);
+            }
+        });
+    },
+
+
+    async playText(text, vc) {
+        // Creates a client
+        const c = new textToSpeech.TextToSpeechClient();
+
+        // Construct the request
+        const request = {
+            input: { text: text },
+            // Select the language and SSML Voice Gender (optional)
+            voice: { languageCode: 'en-US', ssmlGender: 'NEUTRAL' },
+            // Select the type of audio encoding
+            audioConfig: { audioEncoding: 'OGG_OPUS' }
+        };
+
+        // Performs the Text-to-Speech request
+        const [response] = await c.synthesizeSpeech(request);
+        const strm = streamifier.createReadStream(response.audioContent);
+        const audio = strm.pipe(new prism.opus.OggDemuxer());
+
+        let channel = client.channels.find(c => c.id === vc);
+        await channel.join().then(async (connection) => {
+            connection.playOpusStream(audio).on('end', () => {
+                if (queued.length !== 0) {
+                    let next = queued.pop();
+                    module.exports.playText(next.text, next.vc);
+                } else {
+                    channel.leave();
+                }
+            });
+        }).catch(err => console.log(err));
+    }
+};
