--- src/utils/index.js
@@ -1,5 +1,0 @@
-'use strict';
-
-module.exports = {
-	breakSentence: require('./break-sentence'),
-};
